Move PathAndRename class
